Lockscreen charging info: show decimal in battery temperature

+++ packages/SettingsLib/src/com/android/settingslib/fuelgauge/BatteryStatus.java
@@ -55,11 +55,11 @@ public class BatteryStatus {
     public final int maxChargingVoltage;
     public final int maxChargingWattage;
     public final boolean present;
-    public final int temperature;
+    public final float temperature;
 
     public BatteryStatus(int status, int level, int plugged, int health, boolean present,
             int maxChargingCurrent, int maxChargingVoltage,
-            int maxChargingWattage, int temperature) {
+            int maxChargingWattage, float temperature) {
         this.status = status;
         this.level = level;
         this.plugged = plugged;
